+++ src/main/java/io/github/astrapi69/swing/visibility/RenderMode.java
@@ -1,0 +1,14 @@
+package io.github.astrapi69.swing.visibility;
+
+/**
+ * The Enum {@link RenderMode} indicates in which mode a component is
+ */
+public enum RenderMode
+{
+	/** The invisible mode. In this mode the component is not visible */
+	INVISIBLE,
+	/** The viewable mode. In this mode the component is visible but not enabled and not editable */
+	VIEWABLE,
+	/** The editable mode. In this mode the component is visible, enabled and editable */
+	EDITABLE;
+}